health refactor and ui logic creation

+++ Project/Assets/Scripts/Enemies/Interfaces/IDamageable.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void Damage(float damageAmount, Vector2 hitDirection = default);
+    void Damage(int damageAmount, Vector2 hitDirection = default);
 
     void Die();
 
-    float MaxHealth { get; set; }
-    float CurrentHealth { get; set; }
+    int MaxHealth { get; set; }
+    int CurrentHealth { get; set; }
 }
